Наработки по оптимизации (дело близится к завершению)

+++ src/NpoMash.Erm/NpoMash.Erm.Hrm/NpoMash.Erm.Hrm/Optimization/Equality.cs
@@ -4,14 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace NpoMash.Erm.Hrm.Optimization
-{
-    public class Equality: LinearRestriction
-    {
-        private float _EqualTo;
-        public float EqualTo { get { return _EqualTo; } set { _EqualTo = value; } }
-        public override bool Satisfies(ValuesVector values)
-        {
+namespace NpoMash.Erm.Hrm.Optimization {
+    public class Equality : LinearRestriction {
+        private double _EqualTo;
+        public double EqualTo { get { return _EqualTo; } set { _EqualTo = value; } }
+        public override bool Satisfies(ValuesVector values) {
             return Calculate(values) == EqualTo;
         }
     }
