+++ src/NpoMash.Erm/NpoMash.Erm.Hrm/NpoMash.Erm.Hrm/Optimization/Inequality.cs
@@ -11,16 +11,16 @@ namespace NpoMash.Erm.Hrm.Optimization
     /// </summary>
     public class Inequality: LinearRestriction
     {
-        private float _GreaterThan;
+        private double _GreaterThan;
         /// <summary>
         /// Должно быть больше чем это значение
         /// </summary>
-        public float GreaterThan { get { return _GreaterThan; } set { _GreaterThan = value; } }
-        private float _LowerThan;
+        public double GreaterThan { get { return _GreaterThan; } set { _GreaterThan = value; } }
+        private double _LowerThan;
         /// <summary>
         /// Должно быть меньше чем это значение
         /// </summary>
-        public float LowerThan { get { return _LowerThan; } set { _LowerThan = value; } }
+        public double LowerThan { get { return _LowerThan; } set { _LowerThan = value; } }
 
         /// <summary>
         /// Выполняется в заданной точке
@@ -29,7 +29,7 @@ namespace NpoMash.Erm.Hrm.Optimization
         /// <returns></returns>
         public override bool Satisfies(ValuesVector values)
         {
-            float value = Calculate(values);
+            double value = Calculate(values);
             return (value > GreaterThan) && (value < LowerThan);
         }
