+++ src/NpoMash.Erm/NpoMash.Erm.Hrm/NpoMash.Erm.Hrm/Optimization/SimplexRow.cs
@@ -4,26 +4,23 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace NpoMash.Erm.Hrm.Optimization
-{
-    public class SimplexRow
-    {
+namespace NpoMash.Erm.Hrm.Optimization {
+    public class SimplexRow {
         private Variable _BasisVar;
         /// <summary>
         /// Базисная переменная
         /// </summary>
         public Variable BasisVar { get { return _BasisVar; } set { _BasisVar = value; } }
-        private float _VarValue;
+        private double _VarValue;
         /// <summary>
         /// Значение
         /// </summary>
-        public float VarValue { get { return _VarValue; } set { _VarValue = value; } }
+        public double VarValue { get { return _VarValue; } set { _VarValue = value; } }
         /// <summary>
         /// Переменные в строке
         /// </summary>
         public Dictionary<SimplexColumn, SimplexCell> Cells;
-        public SimplexRow(Equality eq, SimplexTable table)
-        {
+        public SimplexRow(Equality eq, SimplexTable table) {
             Cells = new Dictionary<SimplexColumn, SimplexCell>();
             VarValue = eq.EqualTo;
             table.Rows.Add(this);
