+++ src/NpoMash.Erm/NpoMash.Erm.Hrm/NpoMash.Erm.Hrm/Optimization/ValuesVector.cs
@@ -4,9 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace NpoMash.Erm.Hrm.Optimization
-{
-    public class ValuesVector : Dictionary<Variable,float>
-    {
+namespace NpoMash.Erm.Hrm.Optimization {
+    public class ValuesVector : Dictionary<Variable, double> {
     }
 }
